#2947 fix(contacts): fix z-index issue in import overlay. improve colors
I wanted to change button color to primary but it didn't work well with
the split arrow button, and we don't have designs for it, so I skipped
it and just improved text color to prevent disabled look and feel.

I also wanted to replace it with material-ui components, but we decided
not to do it and just quickly fix z-index issue.