#3192 fix(mc): always use template preview image when showing in compose
Ideally, it would be better to always show html instead of preview image because what is
actually sent is the template, not the preview image. But it's more important to use the
same thing no matter it's from "Send to" action in my designs or it's "Edit" action. So
we decided to use preview image for both cases. See issue comments for more info.