wip_fix(email-compose): show image links when editing scheduled emails (#3552)
The problem was in combining different custom block functions. The
way they were combined was wrong. If one of the custom block functions
was returning a non-undefined value, the other ones were not called.
Namely, `signatureCustomBlockFn` was called in this case which was
preventing `blockLevelLinkCustomBlockFn` from being called.
The correct strategy for combining custom block functions is to
merge the result of all of them.