fix(email-compose): prevent MC template preview header layout break (#3709)
PageHeader component ignores input `title` and `backLink` if `location.state.previousPage`
exits, and override them with the ones from `location.state.previousPage`. Even if
they are undefined in `location.state.previousPage`. now it's changed to only override
`title` and `backLink` from `location.state.previousPage` if they are not undefined there.
It may be more reasonable to treat `location.state.previousPage` as a fallback and
use it only if `title` and `backLink` are NOT passed as props. But that would be
a dangerous breaking change that I would avoid.