fix: don't use stale state when moving cursor to beginning (#4243)
An stale editorState from the first render was used to update state
and therefore all decorators that were set by draftjs plugins editor
were overridden by the empty decorator of that stale editor state.

It was gone unnoticed particularly because of the need for ignoring
react hook eslint rule in this case.